[level 0] Title: 배열의 평균값, Time: 0.01 ms, Memory: 4.2 MB -BaekjoonHub

+++ 프로그래머스/0/120817. 배열의 평균값/배열의 평균값.cpp
@@ -4,5 +4,5 @@
 using namespace std;
 
 double solution(vector<int> numbers) {
-    return accumulate(numbers.begin(), numbers.end(), 0) / (float)numbers.size();
+    return accumulate(numbers.begin(), numbers.end(), 0.0) / numbers.size();
 }
